third_party/WebKit/LayoutTests/editing/execCommand/indent-selection.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
        '<span>^Lorem</span><br>',
        'more Lorem!',
        '<blockquote class="webkit-indent-blockquote" style="margin: 0 0 0 40px;">',
            'ipsum<br>',
            '<ul>',
                '<li>Foo</li>',
                '<li>Bar</li>',
                '<li>Baz</li>',
            '</ul>',
            'Dolor',
        '</blockquote>',
        'Sum!<br>',
        '<span>Thing|</span>',
        '</div>',
    ].join(''),
    'indent',
    [
        '<div contenteditable>',
        '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
            '<span>^Lorem<br></span>',
            'more Lorem!',
            '<blockquote class="webkit-indent-blockquote" style="margin: 0 0 0 40px;">',
                'ipsum',
            '</blockquote>',
        '</blockquote>',
        '<blockquote class="webkit-indent-blockquote" style="margin: 0 0 0 40px;">',
            '<ul>',
                '<ul>',
                    '<li>Foo</li>',
                    '<li>Bar</li>',
                    '<li>Baz</li>',
                '</ul>',
            '</ul>',
        '</blockquote>',
        '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
            '<blockquote class="webkit-indent-blockquote" style="margin: 0 0 0 40px;">',
                'Dolor',
            '</blockquote>',
            'Sum!<br>',
            '<span>Thing|</span>',
        '</blockquote>',
        '</div>',
    ].join('')),
    'Indent HTML fragment');
</script>
